+++ AddOns/!KalielsTracker/Modules/Help.lua
@@ -382,6 +382,11 @@ local function SetupTutorials()
 			headingSize = 26,
 			text = "|cff66ff66"..KT.TITLE.." is developed and maintained by one person.|r\n"..ebSpace.."\n"..
 
+					(cTitle.."Version 8.7.1|r\n"..
+					"- ADDED - auto-close Wowhead and YouTube popups after URL copy (see Options > Modules > General)\n"..
+					"- ADDED - remove invalid tracked achievements on start\n"..
+					"\n")..
+
 					(cTitle.."Version 8.7.0|r\n"..
 					"- ADDED - auto expand tracker modules when tracking an item\n"..
 					"- ADDED - support for WoW 12.1.0.69273\n"..